Move gptq trainer init logic to common trainer init
